backend/requirements_production.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9

